feat: Add feedback form functionality to the footer
This commit adds functionality to the footer to display a feedback form when the "Report an Issue" link is clicked. It includes a JavaScript code block in the `_footer.html.erb` file that listens for the click event on the "Report an Issue" link and triggers the display of the feedback form. This change allows users to easily report issues and provide feedback on the website.

Co-authored-by: dependabot[bot] <support@github.com>