feat: Add country_select gem for address form
This commit adds the 'country_select' gem to the Gemfile, enabling the use of a dropdown menu for selecting countries in the address form. This enhancement improves the user experience by providing a more intuitive way to select the country. It also updates the Address model to include associations with home_profiles and campus_profiles, allowing for easier retrieval of profiles associated with a specific address. Additionally, a migration file is created to add home_address_id and campus_address_id references to the profiles table, enabling the association between profiles and addresses. These changes enhance the functionality and organization of the code related to addresses and profiles.

Co-authored-by: dependabot[bot] <support@github.com>