feat: Update `accepts_nested_attributes_for` in Container model
This commit updates the `accepts_nested_attributes_for` method in the `Container` model. It adds the `reject_if: :all_blank` option to the `assignments` association, allowing the rejection of nested attributes if all the attributes are blank. This change ensures that empty assignment records are not saved when updating a container.

Co-authored-by: dependabot[bot] <support@github.com>