feat: Assign container administrator role to creator on container creation
The code changes in the `Container` model ensure that the creator of a container is assigned the role of "Container Administrator" when the container is created. This is achieved by adding an `after_create` callback that creates an assignment for the creator with the appropriate role. This change enhances the user experience by automatically assigning administrative privileges to the creator of a container.

Co-authored-by: dependabot[bot] <support@github.com>