refactor: Remove unnecessary inclusion validation in Category model
The code changes in the Category model remove the unnecessary inclusion validation for the 'kind' attribute. This validation was previously checking for specific values ('Drama', 'Screenplay', 'Non-Fiction', 'Fiction', 'Poetry', 'Novel', 'Short Fiction', 'Text-Image'), but it is now removed. This simplifies the code and improves flexibility for adding new categories in the future.