refactor: Remove CanCan error handling in ProfilesController and UserRolesController
The code changes remove the CanCan error handling in the ProfilesController and UserRolesController. This error handling was redirecting unauthorized users to the root path and returning a forbidden status for JSON requests. However, with the introduction of the pundit and pundit-matchers gems for authorization, the CanCan error handling is no longer necessary. Removing this code simplifies the controllers and ensures consistent authorization handling throughout the application.