refactor: Add file attachment functionality to entry view
The code changes in the _entry.html.erb file add a section to display an attached file for each entry. If an entry has a file attached, a "Download File" link is shown. If no file is attached, a message stating "No file attached" is displayed. This enhancement improves the user experience by allowing users to easily access and download entry files.