refactor: Add "Show" button to contest instances table
This commit adds a "Show" button to each row in the contest instances table on the index page. Clicking the button will navigate to the show page for the corresponding contest instance. This enhancement improves the user experience by providing a convenient way to access detailed information about each contest instance.