Update GitHub Actions workflow to streamline environment variable configuration
This commit modifies the `security.yml` file in the GitHub Actions workflow by removing specific database credentials and replacing them with a more generic `RACK_ENV` variable for the test environment. This change aims to enhance security by minimizing exposure of sensitive information and improving the overall configuration of the CI/CD pipeline.